Fixed test runner. Moved reducer into its own file.
