+++ package.json
@@ -34,6 +34,7 @@
     "lint": "tslint --project tsconfig.json"
   },
   "devDependencies": {
+    "@types/jest": "^23.3.10",
     "@types/moment-duration-format": "^2.2.2",
     "@types/node": "^10.12.11",
     "@types/react": "^16.7.11",
